Listado y detalles de todos los componentes para todos los principals
