+++ src/main/java/acme/roles/Inventor.java
@@ -1,12 +1,16 @@
 package acme.roles;
 
+import java.util.Collection;
+
 import javax.persistence.Entity;
+import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.validation.constraints.NotBlank;
 
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
+import acme.entities.components.Component;
 import acme.entities.patronage.Patronage;
 import acme.framework.roles.UserRole;
 import lombok.Getter;
@@ -40,4 +44,7 @@ public class Inventor extends UserRole {
 
 	@OneToOne
 	Patronage					patronage;
+	
+	@OneToMany
+	Collection<Component>		components;
 }
