+++ src/main/webapp/WEB-INF/views/any/component/list.jsp
@@ -1,0 +1,28 @@
+<%--
+- list.jsp
+-
+- Copyright (C) 2012-2022 Rafael Corchuelo.
+-
+- In keeping with the traditional purpose of furthering education and research, it is
+- the policy of the copyright owner to permit non-commercial use and redistribution of
+- this software. It has been tested carefully, but it is not guaranteed for any particular
+- purposes.  The copyright owner does not offer any warranties or representations, nor do
+- they accept any liabilities with respect to them.
+--%>
+
+<%@page language="java"%>
+
+<%@taglib prefix="jstl" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@taglib prefix="acme" uri="urn:jsptagdir:/WEB-INF/tags"%>
+
+<acme:list readonly="false">
+	<acme:list-column code="any.component.list.label.name" path="name" width="25%"/>
+	<acme:list-column code="any.component.list.label.code" path="code" width="25%"/>
+	<acme:list-column code="any.component.list.label.description" path="description" width="50%"/>
+</acme:list>
+
+<!-- 
+<jstl:if test="${!readonly}">
+	<acme:button code="any.component.list.button.create" action="/any/component/create"/>
+</jstl:if>
+ -->
