+++ src/main/webapp/WEB-INF/views/any/component/messages-en.i18n
@@ -1,0 +1,24 @@
+# i18n-en.i18n
+#
+# Copyright (C) 2012-2022 Rafael Corchuelo.
+#
+# In keeping with the traditional purpose of furthering education and research, it is
+# the policy of the copyright owner to permit non-commercial use and redistribution of
+# this software. It has been tested carefully, but it is not guaranteed for any particular
+# purposes.  The copyright owner does not offer any warranties or representations, nor do
+# they accept any liabilities with respect to them.
+
+any.component.form.label.name = Name:
+any.component.form.label.code = Code:
+any.component.form.label.technology = Technology:
+any.component.form.label.description = Description:
+any.component.form.label.retailPrice = Retail price:
+any.component.form.label.optionalLink = Optional link
+any.component.form.label.inventor = Inventor
+any.component.form.button.create = Create
+ 
+any.component.list.label.name = Name
+any.component.list.label.code = Code
+any.component.list.label.description = Description
+
+any.component.list.button.create = Create component
