+++ src/main/webapp/WEB-INF/views/fragments/menu-es.i18n
@@ -22,12 +22,14 @@ master.menu.anonymous.all-jobs = Ver trabajos
 master.menu.anonymous.list-shouts = Ver gritos
 master.menu.anonymous.list-chirps = Ver chirps
 master.menu.anonymous.list-patronages = Ver mecenazgos
+master.menu.anonymous.list-components = Ver componentes
 
 # master.menu.any #######################################################
 
 master.menu.any = Cualquiera
 master.menu.any.list-chirps = Ver chirps
 master.menu.any.list-patronages = Ver mecenazgos
+master.menu.any.list-components = Ver componentes
 
 # master.menu.authenticated #######################################################
